security(coordinator-api): wallet trust-boundary hardening and exact monetary semantics (B3/B4)
- Close wallet RPC/secret trust-boundary violations and remove client-supplied
  RPC/private-key paths.
- Replace monetary floats with Decimal/atomic units across coordinator wallet,
  marketplace, staking, settlement, and cross-chain models/services.
- Add money.py helpers for exact atomic<->Decimal conversions with float rejection.
- Update affected Alembic migration (e9cf23ae4640) to migrate remaining Float
  monetary columns to Numeric.
- Add property/regression tests for conversion round-trips, fee calculation, and
  adapter float rejection.

+++ apps/coordinator-api/src/coordinator_api/contexts/trading/services/trading_marketplace/dynamic_pricing.py
@@ -243,7 +243,7 @@ class DynamicPricingEngine:
         self,
         resource_id: str,
         resource_type: ResourceType,
-        base_price: float,
+        base_price: Decimal,
         strategy: PricingStrategy | None = None,
         constraints: PriceConstraints | None = None,
         region: str = "global",
