+++ apps/coordinator-api/src/coordinator_api/custom_types.py
@@ -2,6 +2,7 @@
 Shared types and enums for the AITBC Coordinator API
 """
 
+from decimal import Decimal
 from enum import StrEnum
 
 from pydantic import BaseModel
@@ -22,4 +23,4 @@ class Constraints(BaseModel):
     min_vram_gb: int | None = None
     models: list[str] | None = None
     region: str | None = None
-    max_price: float | None = None
+    max_price: Decimal | None = None
